+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by ../iree/build_tools/bazel_to_cmake/bazel_to_cmake.py from   #
-# compiler/src/openxla/compiler/nvgpu/Dialect/Triton/BUILD.bazel               #
+# compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/BUILD.bazel           #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
