+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Conversion/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by ../iree/build_tools/bazel_to_cmake/bazel_to_cmake.py from   #
-# compiler/src/openxla/compiler/nvgpu/Dialect/Triton/IR/test/BUILD.bazel       #
+# compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Conversion/BUILD.bazel#
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -10,14 +10,29 @@
 
 iree_add_all_subdirs()
 
-iree_lit_test_suite(
+iree_cc_library(
   NAME
-    lit
+    TritonToFlowDispatch
+  HDRS
+    "ConvertTritonToFlowDispatch.h"
   SRCS
-    "ops.mlir"
-  TOOLS
-    FileCheck
-    iree-opt
+    "ConvertTritonToFlowDispatch.cpp"
+  DEPS
+    LLVMSupport
+    MLIRIR
+    MLIRPass
+    MLIRTransforms
+    iree::compiler::Dialect::Flow::IR
+    iree::compiler::Dialect::HAL::IR
+    iree::compiler::Utils
+    openxla::compiler::nvgpu::Dialect::TritonFlow::IR
+    # TODO(ezhulenev): Triton dependencies are not part of bazel build.
+    triton::TritonIR
+    triton::TritonTransforms
+    triton::TritonConversion
+    triton::TritonLLVMIR
+    triton::TritonPTX
+  PUBLIC
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
