+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Conversion/ConvertTritonToFlowDispatch.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_CONVERSION_CONVERT_TRITON_TO_CUSTOM_DISPATCH_H_
-#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_CONVERSION_CONVERT_TRITON_TO_CUSTOM_DISPATCH_H_
+#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_CONVERSION_CONVERT_TRITON_TO_FLOW_DISPATCH_H_
+#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_CONVERSION_CONVERT_TRITON_TO_FLOW_DISPATCH_H_
 
 #include "iree/compiler/Utils/OptionUtils.h"
 #include "mlir/IR/PatternMatch.h"
@@ -49,4 +49,4 @@ void buildTritonCompilationPipeline(mlir::OpPassManager &pm,
 
 }  // namespace openxla::compiler::nvgpu::triton
 
-#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_CONVERSION_CONVERT_TRITON_TO_CUSTOM_DISPATCH_H_
+#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_CONVERSION_CONVERT_TRITON_TO_FLOW_DISPATCH_H_
