+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/IR/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by ../iree/build_tools/bazel_to_cmake/bazel_to_cmake.py from   #
-# compiler/src/openxla/compiler/nvgpu/Dialect/Triton/Conversion/BUILD.bazel    #
+# compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/IR/BUILD.bazel        #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,27 +12,37 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    TritonToFlowDispatch
+    IR
   HDRS
-    "ConvertTritonToFlowDispatch.h"
+    "TritonFlowDialect.h"
+    "TritonFlowOps.h"
+  TEXTUAL_HDRS
+    "TritonFlowDialect.cpp.inc"
+    "TritonFlowDialect.h.inc"
+    "TritonFlowOps.cpp.inc"
+    "TritonFlowOps.h.inc"
   SRCS
-    "ConvertTritonToFlowDispatch.cpp"
+    "TritonFlowDialect.cpp"
+    "TritonFlowOps.cpp"
   DEPS
+    ::TritonFlowOpsGen
     LLVMSupport
     MLIRIR
-    MLIRPass
-    MLIRTransforms
-    iree::compiler::Dialect::Flow::IR
-    iree::compiler::Dialect::HAL::IR
-    iree::compiler::Utils
-    openxla::compiler::nvgpu::Dialect::Triton::IR
-    # TODO(ezhulenev): Triton dependencies are not part of bazel build.
-    triton::TritonIR
-    triton::TritonTransforms
-    triton::TritonConversion
-    triton::TritonLLVMIR
-    triton::TritonPTX
+    MLIRSupport
+    openxla::compiler::nvgpu::defs
   PUBLIC
+)
+
+iree_tablegen_library(
+  NAME
+    TritonFlowOpsGen
+  TD_FILE
+    "TritonFlowOps.td"
+  OUTS
+    --gen-dialect-decls --dialect=triton TritonFlowDialect.h.inc
+    --gen-dialect-defs --dialect=triton TritonFlowDialect.cpp.inc
+    --gen-op-decls TritonFlowOps.h.inc
+    --gen-op-defs TritonFlowOps.cpp.inc
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
