+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/IR/TritonFlowDialect.h
@@ -4,13 +4,13 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_DIALECT_H
-#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_DIALECT_H
+#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_DIALECT_H
+#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_DIALECT_H
 
 #include "mlir/IR/Dialect.h"
 #include "mlir/IR/DialectImplementation.h"
 
 // Must be included after all MLIR headers.
-#include "openxla/compiler/nvgpu/Dialect/Triton/IR/TritonDialect.h.inc"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/IR/TritonFlowDialect.h.inc"
 
-#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_DIALECT_H
+#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_DIALECT_H
