+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/IR/TritonFlowOps.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_OPS_H
-#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_OPS_H
+#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_OPS_H
+#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_OPS_H
 
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Dialect.h"
@@ -17,6 +17,6 @@
 #include "mlir/Interfaces/SideEffectInterfaces.h"
 
 #define GET_OP_CLASSES
-#include "openxla/compiler/nvgpu/Dialect/Triton/IR/TritonOps.h.inc"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/IR/TritonFlowOps.h.inc"
 
-#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_IR_TRITON_OPS_H
+#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_IR_TRITONFLOW_OPS_H
