+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by ../iree/build_tools/bazel_to_cmake/bazel_to_cmake.py from   #
-# compiler/src/openxla/compiler/nvgpu/Dialect/Triton/IR/test/BUILD.bazel       #
+# compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/BUILD.bazel#
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -10,14 +10,42 @@
 
 iree_add_all_subdirs()
 
-iree_lit_test_suite(
+iree_tablegen_library(
   NAME
-    lit
+    PassesIncGen
+  TD_FILE
+    "Passes.td"
+  OUTS
+    --gen-pass-decls Passes.h.inc
+)
+
+iree_cc_library(
+  NAME
+    PassHeaders
+  HDRS
+    "PassDetail.h"
+    "Passes.h"
+    "Passes.h.inc"
+  DEPS
+    ::PassesIncGen
+    MLIRPass
+    MLIRTransforms
+  PUBLIC
+)
+
+iree_cc_library(
+  NAME
+    Transforms
+  HDRS
+    "Passes.h"
   SRCS
-    "triton_to_flow_dispatch.mlir"
-  TOOLS
-    FileCheck
-    iree-opt
+    "ConvertTritonToFlowDispatch.cpp"
+    "Passes.cpp"
+  DEPS
+    ::PassHeaders
+    MLIRPass
+    openxla::compiler::nvgpu::Dialect::TritonFlow::Conversion::TritonToFlowDispatch
+  PUBLIC
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
