+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/ConvertTritonToFlowDispatch.cpp
@@ -4,20 +4,21 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "openxla/compiler/nvgpu/Dialect/Triton/Conversion/ConvertTritonToFlowDispatch.h"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/Conversion/ConvertTritonToFlowDispatch.h"
 
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/IR/Types.h"
 #include "mlir/Transforms/DialectConversion.h"
-#include "openxla/compiler/nvgpu/Dialect/Triton/Transforms/PassDetail.h"
-#include "openxla/compiler/nvgpu/Dialect/Triton/Transforms/Passes.h"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/PassDetail.h"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/Passes.h"
 
 #define GEN_PASS_DEF_CONVERTTRITONTOFLOWDISPATCH
-#include "openxla/compiler/nvgpu/Dialect/Triton/Transforms/Passes.h.inc"
+#include "openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/Passes.h.inc"
 
 namespace IREE = mlir::iree_compiler::IREE;
 
 using namespace mlir;
+using namespace mlir::triton;
 
 namespace openxla::compiler::nvgpu::triton {
 
@@ -31,9 +32,9 @@ class ConvertTritonToFlowDispatch
     TypeConverter typeConverter;
     typeConverter.addConversion([](Type type) { return type; });
 
-    // Ensure all cuDNN dialect operations go away.
+    // Ensure all TritonFlow operations go away.
     ConversionTarget conversionTarget(*context);
-    conversionTarget.addIllegalDialect<triton::TritonDialect>();
+    conversionTarget.addIllegalDialect<TritonFlowDialect>();
     conversionTarget.addLegalDialect<IREE::HAL::HALDialect>();
     conversionTarget.addLegalDialect<IREE::Flow::FlowDialect>();
