+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/Passes.h
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_TRANSFORMS_PASSES_H_
-#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_TRANSFORMS_PASSES_H_
+#ifndef OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_TRANSFORMS_PASSES_H_
+#define OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_TRANSFORMS_PASSES_H_
 
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/Pass/Pass.h"
@@ -30,4 +30,4 @@ void registerOpenXlaTritonPipelines();
 
 }  // namespace openxla::compiler::nvgpu::triton
 
-#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITON_TRANSFORMS_PASSES_H_
+#endif  // OPENXLA_COMPILER_NVGPU_DIALECT_TRITONFLOW_TRANSFORMS_PASSES_H_
