+++ compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/Transforms/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by ../iree/build_tools/bazel_to_cmake/bazel_to_cmake.py from   #
-# compiler/src/openxla/compiler/nvgpu/Dialect/Triton/IR/BUILD.bazel            #
+# compiler/src/openxla/compiler/nvgpu/Dialect/TritonFlow/IR/test/BUILD.bazel   #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -10,39 +10,14 @@
 
 iree_add_all_subdirs()
 
-iree_cc_library(
+iree_lit_test_suite(
   NAME
-    IR
-  HDRS
-    "TritonDialect.h"
-    "TritonOps.h"
-  TEXTUAL_HDRS
-    "TritonDialect.cpp.inc"
-    "TritonDialect.h.inc"
-    "TritonOps.cpp.inc"
-    "TritonOps.h.inc"
+    lit
   SRCS
-    "TritonDialect.cpp"
-    "TritonOps.cpp"
-  DEPS
-    ::TritonOpsGen
-    LLVMSupport
-    MLIRIR
-    MLIRSupport
-    openxla::compiler::nvgpu::defs
-  PUBLIC
-)
-
-iree_tablegen_library(
-  NAME
-    TritonOpsGen
-  TD_FILE
-    "TritonOps.td"
-  OUTS
-    --gen-dialect-decls --dialect=triton TritonDialect.h.inc
-    --gen-dialect-defs --dialect=triton TritonDialect.cpp.inc
-    --gen-op-decls TritonOps.h.inc
-    --gen-op-defs TritonOps.cpp.inc
+    "triton_to_flow_dispatch.mlir"
+  TOOLS
+    FileCheck
+    iree-opt
 )
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
